# Optimize Cartridge::Mbc1